# docs/swagger.yaml
basePath: /
definitions:
  handlers.ProbeConfig:
    properties:
      bootDelaySeconds:
        description: |-
          Startup only: wall-clock seconds from process start before a success is allowed.
          example: 10
        example: 10
        type: number
      delaySeconds:
        description: |-
          sleep before answering (seconds)
          example: 0
        example: 0
        type: number
      flapEvery:
        description: |-
          Flap: every Nth request fails (0 = use time-based flapSeconds).
          example: 2
        example: 2
        type: integer
      flapSeconds:
        description: |-
          Flap (mode=flap): half-period seconds (default 5).
          example: 5
        example: 5
        type: number
      mode:
        description: |-
          ok | fail | delay | flap
          example: fail
        example: fail
        type: string
    type: object
  handlers.ProbeSnapshot:
    properties:
      live:
        $ref: '#/definitions/handlers.ProbeConfig'
      liveFlapPhase:
        description: Which half of a time-based flap we are in right now (ok|fail|n/a).
        type: string
      ready:
        $ref: '#/definitions/handlers.ProbeConfig'
      readyFlapPhase:
        type: string
      startup:
        $ref: '#/definitions/handlers.ProbeConfig'
      startupFlapPhase:
        type: string
      startupLatched:
        type: boolean
      startupWouldPass:
        type: boolean
      uptimeSeconds:
        type: number
    type: object
  handlers.ProbeUpdate:
    properties:
      live:
        $ref: '#/definitions/handlers.ProbeConfig'
      ready:
        $ref: '#/definitions/handlers.ProbeConfig'
      resetStartupLatch:
        description: 'example: true'
        example: true
        type: boolean
      startup:
        $ref: '#/definitions/handlers.ProbeConfig'
    type: object
  handlers.ProxyRequest:
    properties:
      body:
        description: |-
          Optional body (string; use for JSON text, form, etc.)
          example: {"hello":"cluster"}
        example: '{"hello":"cluster"}'
        type: string
      forwardIncomingHeaders:
        description: |-
          When true (default), copy inbound request headers onto the outbound call
          (minus hop-by-hop). Tracing headers like X-Request-Id ride along.
          example: true
        example: true
        type: boolean
      forwardSensitiveHeaders:
        description: |-
          When true, also forward Authorization / Cookie from the inbound request.
          Default false so your bearer token for *this* api is not sent east-west by accident.
          example: false
        example: false
        type: boolean
      headers:
        additionalProperties:
          type: string
        description: |-
          Extra headers to set/override on the outbound request
          example: {"X-Demo":"from-swagger"}
        example:
          X-Demo: from-swagger
        type: object
      method:
        description: |-
          HTTP method (default GET)
          example: GET
        example: GET
        type: string
      raw:
        description: |-
          When true, return the upstream body as the raw HTTP response (status + headers from upstream).
          Default false → JSON wrap with request/response/meta (includes upstream headers + body).
          example: false
        example: false
        type: boolean
      timeoutSeconds:
        description: |-
          Timeout for the outbound call (default 10; capped by MAX_DELAY_SECONDS)
          example: 15
        example: 15
        type: number
      url:
        description: |-
          Absolute URL to call (demo: httpbin echoes method/headers back as JSON)
          example: https://httpbin.org/get
        example: https://httpbin.org/get
        type: string
    required:
    - url
    type: object
info:
  contact: {}
  description: |-
    HTTP sidekick for cluster-utils — drop into a namespace and poke probes, ingress headers, env, and east-west hops.

    • Authorize (top-right lock): Bearer &lt;token&gt;  e.g. local podman → Bearer dev
    • Probes: /startupz /livez /readyz  ·  Control: GET|PUT /a/control/probes
    • Proxy: GET /a/proxy?url=https://httpbin.org/get  (or POST JSON body)
    • Pair: https://github.com/donkeyx/cluster-utils
  title: Cluster Utils API
  version: 2.5.0
paths:
  /a/control/probes:
    get:
      description: Current live/ready/startup config + startup latch + flap phase
        + uptime
      operationId: getProbes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProbeSnapshot'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get probe control state
    put:
      consumes:
      - application/json
      description: 'Partial update of live/ready/startup. Example: fail readiness
        so the pod drops from Service endpoints (no restart). resetStartupLatch re-runs
        cold start without restarting the process. Auth: Authorize lock with "Bearer
        dev".'
      operationId: putProbes
      parameters:
      - description: Partial update — schema examples show fail-readiness shape
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ProbeUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProbeSnapshot'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update probe control state
  /a/env:
    get:
      description: 'Env dump so you can check secrets/configmaps/task params actually
        landed. Behind auth under /a/. Use the Authorize button (value: Bearer &lt;token&gt;)
        — do not leave a separate Authorization param empty.'
      operationId: env
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get environment variables
  /a/proxy:
    get:
      description: 'Auth: top-right Authorize with "Bearer dev". Simple query-only
        hop (no body — browsers forbid GET+body). Default url hits httpbin so you
        see method/headers echoed. Prefer POST /a/proxy for full JSON control.'
      operationId: proxyGet
      parameters:
      - default: https://httpbin.org/get
        description: absolute URL to fetch
        example: https://httpbin.org/get
        in: query
        name: url
        type: string
      - default: GET
        description: HTTP method for the outbound call
        example: GET
        in: query
        name: method
        type: string
      - default: 15
        description: outbound timeout seconds
        example: 15
        in: query
        name: timeoutSeconds
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Proxy GET hop (auth)
    post:
      consumes:
      - application/json
      description: 'Auth: top-right Authorize with "Bearer dev". Full JSON control
        — example body calls https://httpbin.org/get so the wrap shows upstream status/headers/body.
        North→south then east-west; SSRF if left unauthenticated.'
      operationId: proxyPost
      parameters:
      - description: proxy request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ProxyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Proxy POST hop (auth)
...
  /delay/{seconds}:
    get:
      description: Sleep N seconds then return 200. Cap is MAX_DELAY_SECONDS env (default
        120, hard max 600). Try 2 for a slow upstream demo; use probe delaySeconds
        to trip kube timeoutSeconds.
      operationId: delay
      parameters:
      - default: 2
        description: seconds to sleep
        example: 2
        in: path
        name: seconds
        required: true
        type: number
      produces:
      - text/plain
      responses:
        "200":
          description: delayed
          schema:
            type: string
      summary: Delay then OK
  /echo:
    delete:
      consumes:
      - text/plain
      description: Bounce method, path, query, headers and body back as json. Try
        POST with any body and custom headers (e.g. X-Request-Id) to see what arrived.
      operationId: echo
      parameters:
      - default: '{"hello":"from-swagger"}'
        description: optional body to echo
        example: '{"hello":"from-swagger"}'
        in: body
        name: body
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Echo request
    get:
      consumes:
      - text/plain
      description: Bounce method, path, query, headers and body back as json. Try
        POST with any body and custom headers (e.g. X-Request-Id) to see what arrived.
      operationId: echo
      parameters:
      - default: '{"hello":"from-swagger"}'
        description: optional body to echo
        example: '{"hello":"from-swagger"}'
        in: body
        name: body
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Echo request
    patch:
      consumes:
      - text/plain
      description: Bounce method, path, query, headers and body back as json. Try
        POST with any body and custom headers (e.g. X-Request-Id) to see what arrived.
      operationId: echo
      parameters:
      - default: '{"hello":"from-swagger"}'
        description: optional body to echo
        example: '{"hello":"from-swagger"}'
        in: body
        name: body
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Echo request
    post:
      consumes:
      - text/plain
      description: Bounce method, path, query, headers and body back as json. Try
        POST with any body and custom headers (e.g. X-Request-Id) to see what arrived.
      operationId: echo
      parameters:
      - default: '{"hello":"from-swagger"}'
        description: optional body to echo
        example: '{"hello":"from-swagger"}'
        in: body
        name: body
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Echo request
    put:
      consumes:
      - text/plain
      description: Bounce method, path, query, headers and body back as json. Try
        POST with any body and custom headers (e.g. X-Request-Id) to see what arrived.
      operationId: echo
      parameters:
      - default: '{"hello":"from-swagger"}'
        description: optional body to echo
        example: '{"hello":"from-swagger"}'
        in: body
        name: body
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Echo request
...
  /startupz:
    get:
      description: 'Cold start latch: 503 until bootDelay, then sticky 200. mode=fail
        never latches. mode=flap oscillates until a success latches (or forever if
        you only hit fail phases — use flap carefully here).'
      operationId: startupz
      produces:
      - text/plain
      responses:
        "200":
          description: started
          schema:
            type: string
        "503":
          description: starting
          schema:
            type: string
      summary: Startup (startupz)
  /status/{code}:
    get:
      description: Respond with whatever http status you pass (100-599). Try 418 (teapot),
        503, or 502 for ingress/retry tests.
      operationId: status
      parameters:
      - default: 418
        description: HTTP status code
        example: 418
        in: path
        name: code
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: status body
          schema:
            type: string
      summary: Fixed status code
...
securityDefinitions:
  BearerAuth:
    description: 'Top-right Authorize lock only. Value MUST be: Bearer <token>  e.g.
      Bearer dev'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"