# routes/routes.go
package routes

import (
	"github.com/donkeyx/cluster-utils-api/docs"
	"github.com/donkeyx/cluster-utils-api/handlers"
	"github.com/donkeyx/cluster-utils-api/middleware"
	"net/http"
	"strings"
	"sync"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// swaggerInfoMu guards docs.SwaggerInfo Host/Schemes when serving the UI for different origins.
var swaggerInfoMu sync.Mutex

func SetupRouter(logger *zap.Logger, st string, r *gin.Engine) {
	// Middleware (otel / metrics / log / recover) is registered in main before this.

	// Redirect to swagger docs
	r.GET("/", func(c *gin.Context) {
		c.Redirect(http.StatusFound, "/api-docs/index.html")
	})

	// Swagger UI: persist Authorize token in the browser; host/scheme follow where you opened
	// the page (so port-forward / docker / cluster DNS all work). Optional overrides:
	//   /api-docs/index.html?host=my-svc:8080&scheme=http
	r.GET("/api-docs/*any", swaggerHandler())

	r.GET("/help", handlers.HelpHandler)
	r.GET("/version", handlers.VersionHandler)
	r.GET("/metrics", handlers.PrometheusMetricsHandler())

	// Kube-style probes (plus older aliases)
	// live  = liveness  → restart on fail
	// ready = readiness → leave Service endpoints on fail
	// startup = cold start latch → kube only until first success
	r.GET("/livez", handlers.LiveHandler)
	r.GET("/healthz", handlers.HealthzHandler)
	r.GET("/health", handlers.HealthHandler)

	r.GET("/readyz", handlers.ReadyzHandler)
	r.GET("/ready", handlers.ReadyHandler)

	r.GET("/startupz", handlers.StartupHandler)
	r.GET("/startup", handlers.StartupHandler)

	r.GET("/headers", handlers.HeadersHandler)
	r.GET("/debug", handlers.DebugHandler)
	r.GET("/ping", handlers.PingHandler)

	r.GET("/status/:code", handlers.StatusHandler)
	r.GET("/delay/:seconds", handlers.DelayHandler)
	r.Any("/echo", handlers.EchoHandler)

	// Sensitive / abusable — bearer auth required (see README security section)
	authGroup := r.Group("/a")
	authGroup.Use(middleware.AuthMiddleware(logger, st))
	authGroup.GET("/env", handlers.EnvHandler)
	authGroup.GET("/control/probes", handlers.GetProbesHandler)
	authGroup.PUT("/control/probes", handlers.PutProbesHandler)
	// open /proxy would be SSRF (scan cluster, hit metadata, etc.)
	// Separate handlers so Swagger GET has no body (browsers reject GET+body).
	authGroup.GET("/proxy", handlers.ProxyGetHandler)
	authGroup.POST("/proxy", handlers.ProxyPostHandler)
}

func swaggerHandler() gin.HandlerFunc {
	// Empty host in the generated spec would also work; we set Host from the request
	// so the Swagger top bar shows a real target and Try it out hits the right place.
	handler := ginSwagger.WrapHandler(
		swaggerFiles.Handler,
		ginSwagger.PersistAuthorization(true),
		ginSwagger.DefaultModelsExpandDepth(-1),
	)

	return func(c *gin.Context) {
		host := strings.TrimSpace(c.Query("host"))
		if host == "" {
			host = c.Request.Host
		}

		scheme := strings.ToLower(strings.TrimSpace(c.Query("scheme")))
		if scheme != "http" && scheme != "https" {
			if c.Request.TLS != nil || c.GetHeader("X-Forwarded-Proto") == "https" {
				scheme = "https"
			} else {
				scheme = "http"
			}
		}

		// Serialize updates to the global SwaggerInfo used when doc.json is generated.
		swaggerInfoMu.Lock()
		docs.SwaggerInfo.Host = host
		docs.SwaggerInfo.Schemes = []string{scheme}
		docs.SwaggerInfo.BasePath = "/"
		handler(c)
		swaggerInfoMu.Unlock()
	}
}